+++ test/test_ai_transcripts.py
@@ -9,7 +9,7 @@ from openai import OpenAI
 class TestAITranscripts(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        cls.ticker = Ticker("AMD", http_proxy="http://127.0.0.1:33210", log_level=logging.DEBUG)
+        cls.ticker = Ticker("NVDA", http_proxy="http://127.0.0.1:33210", log_level=logging.DEBUG)
 
         key = Path(__file__).parent.joinpath("siliconflow_api.key").read_text(encoding="utf-8")
         cls.llm = OpenAI(
@@ -29,4 +29,13 @@ class TestAITranscripts(unittest.TestCase):
             2,
             self.llm,
             OpenAIConfiguration(model='Qwen/Qwen3-8B', temperature=0))
+        print(res.to_string())
+
+    def test_analyze_financial_metrics_change_for_this_quarter_with_ai(self):
+        transcripts = self.ticker.earning_call_transcripts()
+        res = transcripts.analyze_financial_metrics_change_for_this_quarter_with_ai(
+            2026,
+            2,
+            self.llm,
+            OpenAIConfiguration(model='Qwen/Qwen3-8B', temperature=0, top_p=0.99, tool_choice="auto"))
         print(res.to_string())